+++ src/test/java/com/iexec/resultproxy/ipfs/IpfsResultServiceTest.java
@@ -1,21 +1,52 @@
+/*
+ * Copyright 2020-2024 IEXEC BLOCKCHAIN TECH
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.iexec.resultproxy.ipfs;
 
-import com.iexec.resultproxy.chain.IexecHubService;
-
+import com.iexec.commons.poco.utils.BytesUtils;
+import com.iexec.resultproxy.ipfs.task.IpfsNameService;
+import com.iexec.resultproxy.result.Result;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
+import org.mockito.Spy;
 
+import java.util.Optional;
+import java.util.stream.Stream;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.Mockito.doReturn;
+import static org.mockito.Mockito.when;
 
 class IpfsResultServiceTest {
 
     @Mock
-    private IexecHubService iexecHubService;
+    private IpfsNameService ipfsNameService;
 
     @Mock
     private IpfsService ipfsService;
 
+    @Spy
     @InjectMocks
     private IpfsResultService ipfsResultService;
 
@@ -27,46 +58,59 @@ class IpfsResultServiceTest {
         chainTaskId = "0x1";
     }
 
-    /*
+    // region addResult
     @Test
-    void shouldGetIpfsHashFromChainTaskId() {
+    void shouldNotAddResult() {
+        when(ipfsNameService.getIpfsHashForTask(anyString())).thenReturn("QmfZ88JXmx2FJsAxT4ZsJBVhBUXdPoRbDZhbkSS1WsMbUA");
+        assertThat(ipfsResultService.addResult(Result.builder().chainTaskId(chainTaskId).build(), new byte[0])).isEmpty();
+    }
+
+    @Test
+    void shouldAddResult() {
+        when(ipfsNameService.getIpfsHashForTask(anyString())).thenReturn("");
+        when(ipfsService.add(any(), any())).thenReturn("QmfZ88JXmx2FJsAxT4ZsJBVhBUXdPoRbDZhbkSS1WsMbUA");
+        assertThat(ipfsResultService.addResult(Result.builder().chainTaskId(chainTaskId).build(), new byte[0]))
+                .isEqualTo("/ipfs/QmfZ88JXmx2FJsAxT4ZsJBVhBUXdPoRbDZhbkSS1WsMbUA");
+    }
+    // endregion
+
+    // region doesResultExist
+    @Test
+    void shouldResultExist() {
+        doReturn(Optional.of(new byte[0])).when(ipfsResultService).getResult(chainTaskId);
+        assertThat(ipfsResultService.doesResultExist(chainTaskId)).isTrue();
+    }
+
+    @Test
+    void shouldResultNotExist() {
+        doReturn(Optional.empty()).when(ipfsResultService).getResult(chainTaskId);
+        assertThat(ipfsResultService.doesResultExist(chainTaskId)).isFalse();
+    }
+    // endregion
+
+    // region getResult
+    @Test
+    void shouldGetResult() {
         String resultLink = "/ipfs/QmfZ88JXmx2FJsAxT4ZsJBVhBUXdPoRbDZhbkSS1WsMbUA";
-        when(iexecHubService.getTaskResults(chainTaskId, 0)).thenReturn(BytesUtils.bytesToString(resultLink.getBytes()));
-
-        assertThat(ipfsResultService.getIpfsHashFromChainTaskId(chainTaskId).equals("QmfZ88JXmx2FJsAxT4ZsJBVhBUXdPoRbDZhbkSS1WsMbUA")).isTrue();
+        when(ipfsNameService.getIpfsHashForTask(chainTaskId)).thenReturn(BytesUtils.bytesToString(resultLink.getBytes()));
+        assertThat(ipfsResultService.getResult(chainTaskId)).isNotNull();
     }
 
-    @Test
-    void shouldNotGetIpfsHashFromChainTaskIdSinceNoTaskResult() {
-        String resultLink = "/";
-        when(iexecHubService.getTaskResults(chainTaskId, 0)).thenReturn(BytesUtils.bytesToString(resultLink.getBytes()));
-
-        assertThat(ipfsResultService.getIpfsHashFromChainTaskId(chainTaskId)).isEmpty();
+    @ParameterizedTest
+    @MethodSource(value = "provideIpfsHash")
+    void shouldNotGetResultWhenBadIpfsHash(String ipfsHash) {
+        when(ipfsNameService.getIpfsHashForTask(chainTaskId)).thenReturn(ipfsHash);
+        assertThat(ipfsResultService.getResult(chainTaskId)).isEmpty();
     }
 
-    @Test
-    void shouldNotGetIpfsHashFromChainTaskIdSinceNotHexaString() {
-        when(iexecHubService.getTaskResults(chainTaskId, 0)).thenReturn("0xefg");
-
-        assertThat(ipfsResultService.getIpfsHashFromChainTaskId(chainTaskId)).isEmpty();
+    static Stream<Arguments> provideIpfsHash() {
+        return Stream.of(
+                Arguments.of("0xefg"),
+                Arguments.of(BytesUtils.bytesToString("/".getBytes())),
+                Arguments.of(BytesUtils.bytesToString("/ipfs/ipfs/123".getBytes())),
+                Arguments.of(BytesUtils.bytesToString("https://customrepo.com/results/abc".getBytes()))
+        );
     }
-
-    @Test
-    void shouldNotGetIpfsHashFromChainTaskIdSinceNotIpfsLink() {
-        String resultLink = "https://customrepo.com/results/abc";
-        when(iexecHubService.getTaskResults(chainTaskId, 0)).thenReturn(BytesUtils.bytesToString(resultLink.getBytes()));
-
-        assertThat(ipfsResultService.getIpfsHashFromChainTaskId(chainTaskId)).isEmpty();
-    }
-
-    @Test
-    void shouldNotGetIpfsHashFromChainTaskIdSinceNotIpfsHash() {
-        String resultLink = "/ipfs/ipfs/123";
-        when(iexecHubService.getTaskResults(chainTaskId, 0)).thenReturn(BytesUtils.bytesToString(resultLink.getBytes()));
-
-        assertThat(ipfsResultService.getIpfsHashFromChainTaskId(chainTaskId)).isEmpty();
-    }
-
-    */
+    // endregion
 
 }